LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-pseudo/text-selection-expected.txt

helloworld
helloworld
helloworld

FAIL Selection ending in ::before assert_equals: toString expected "hello" but got ""
FAIL Selection contained in ::before assert_equals: anchorOffset expected 0 but got 5
FAIL Selection ending in ::marker assert_equals: toString expected "hello" but got ""
PASS Selection contained in ::marker
FAIL Selection ending in ::before-marker assert_equals: toString expected "hello" but got ""
FAIL Selection contained in ::before-marker assert_equals: anchorOffset expected 0 but got 5
